--- tests/test_configs/with_sha256_hash/config.co
@@ -1,12 +1,0 @@
-define user ask capabilities
-  "What can you do?"
-  "What can you help me with?"
-  "tell me what you can do"
-  "tell me about you"
-
-define bot inform capabilities
-  "I am an AI assistant that helps answer questions."
-
-define flow
-  user ask capabilities
-  bot inform capabilities
